refresh logic and warnings fixes
