+++ novawallet/Modules/Banners/BannersViewController+UICollectionView.swift
@@ -1,0 +1,84 @@
+import Foundation
+import UIKit
+
+// MARK: UICollectionViewDataSource
+
+extension BannersViewController: UICollectionViewDataSource {
+    func collectionView(
+        _: UICollectionView,
+        numberOfItemsInSection _: Int
+    ) -> Int {
+        if let loopedViewModels {
+            loopedViewModels.count
+        } else if let viewModels {
+            viewModels.count
+        } else {
+            0
+        }
+    }
+
+    func collectionView(
+        _ collectionView: UICollectionView,
+        cellForItemAt indexPath: IndexPath
+    ) -> UICollectionViewCell {
+        guard let viewModels else {
+            return UICollectionViewCell()
+        }
+
+        let cell = collectionView.dequeueReusableCellWithType(
+            BannerCollectionViewCell.self,
+            for: indexPath
+        )!
+
+        let actualViewModels = if let loopedViewModels {
+            loopedViewModels
+        } else {
+            viewModels
+        }
+
+        cell.bind(with: actualViewModels[indexPath.item])
+
+        return cell
+    }
+}
+
+// MARK: UICollectionViewDelegate
+
+extension BannersViewController: UICollectionViewDelegate {
+    func collectionView(
+        _: UICollectionView,
+        didSelectItemAt indexPath: IndexPath
+    ) {
+        guard let viewModels else { return }
+
+        let index = indexPath.item
+
+        let bannerId = if let loopedViewModels {
+            loopedViewModels[index].id
+        } else {
+            viewModels[index].id
+        }
+
+        presenter.action(for: bannerId)
+    }
+}
+
+// MARK: UICollectionViewDelegateFlowLayout
+
+extension BannersViewController: UICollectionViewDelegateFlowLayout {
+    func collectionView(
+        _: UICollectionView,
+        layout _: UICollectionViewLayout,
+        sizeForItemAt _: IndexPath
+    ) -> CGSize {
+        rootView.backgroundView.bounds.size
+    }
+
+    func collectionView(
+        _: UICollectionView,
+        layout _: UICollectionViewLayout,
+        minimumLineSpacingForSectionAt _: Int
+    ) -> CGFloat {
+        .zero
+    }
+}
